Refactor Point nD test cases
