+++ packages/core/src/common/point-nd/test-util.ts
@@ -1,0 +1,34 @@
+import {findConstraint} from '../constraint/composite';
+import {Constraint} from '../constraint/constraint';
+import {DefiniteRangeConstraint} from '../constraint/definite-range';
+import {RangeConstraint} from '../constraint/range';
+import {StepConstraint} from '../constraint/step';
+
+/**
+ * Finds a range from number constraint.
+ * @param c The number constraint.
+ * @return A list that contains a minimum value and a max value.
+ */
+function findNumberRange(
+	c: Constraint<number>,
+): [number | undefined, number | undefined] {
+	const drc = findConstraint(c, DefiniteRangeConstraint);
+	if (drc) {
+		return [drc.values.get('min'), drc.values.get('max')];
+	}
+	const rc = findConstraint(c, RangeConstraint);
+	if (rc) {
+		return [rc.values.get('min'), rc.values.get('max')];
+	}
+	return [undefined, undefined];
+}
+
+export function getDimensionProps(c: Constraint<number>) {
+	const [min, max] = findNumberRange(c);
+	const sc = findConstraint(c, StepConstraint);
+	return {
+		max: max,
+		min: min,
+		step: sc?.step,
+	};
+}
