+++ packages/core/src/input-binding/point-2d/plugin-test.ts
@@ -4,13 +4,11 @@ import {describe, it} from 'mocha';
 import {InputBindingController} from '../../blade/binding/controller/input-binding';
 import {BindingTarget} from '../../common/binding/target';
 import {InputBindingValue} from '../../common/binding/value/input-binding';
-import {findConstraint} from '../../common/constraint/composite';
 import {Constraint} from '../../common/constraint/constraint';
-import {StepConstraint} from '../../common/constraint/step';
 import {ComplexValue} from '../../common/model/complex-value';
 import {getBoundValue} from '../../common/model/test-util';
 import {Value} from '../../common/model/value';
-import {findNumberRange} from '../../common/number/util';
+import {getDimensionProps} from '../../common/point-nd/test-util';
 import {createTestWindow} from '../../misc/dom-test-util';
 import {Tuple2} from '../../misc/type-util';
 import {PointNdConstraint} from '../common/constraint/point-nd';
@@ -24,16 +22,6 @@ function getPoint2dConstraint(
 ): PointNdConstraint<Point2d> {
 	return (getBoundValue(v) as ComplexValue<unknown>)
 		.constraint as PointNdConstraint<Point2d>;
-}
-
-function getDimensionProps(c: Constraint<number>) {
-	const [min, max] = findNumberRange(c);
-	const sc = findConstraint(c, StepConstraint);
-	return {
-		max: max,
-		min: min,
-		step: sc?.step,
-	};
 }
 
 describe(getSuitableMax.name, () => {
