+++ packages/core/src/input-binding/point-3d/plugin-test.ts
@@ -9,7 +9,7 @@ import {Constraint} from '../../common/constraint/constraint';
 import {StepConstraint} from '../../common/constraint/step';
 import {ComplexValue} from '../../common/model/complex-value';
 import {getBoundValue} from '../../common/model/test-util';
-import {findNumberRange} from '../../common/number/util';
+import {getDimensionProps} from '../../common/point-nd/test-util';
 import {createTestWindow} from '../../misc/dom-test-util';
 import {Tuple3} from '../../misc/type-util';
 import {PointNdConstraint} from '../common/constraint/point-nd';
@@ -23,16 +23,6 @@ function getPoint3dConstraint(
 ): PointNdConstraint<Point3d> {
 	return (getBoundValue(v) as ComplexValue<unknown>)
 		.constraint as PointNdConstraint<Point3d>;
-}
-
-function getDimensionProps(c: Constraint<number>) {
-	const [min, max] = findNumberRange(c);
-	const sc = findConstraint(c, StepConstraint);
-	return {
-		max: max,
-		min: min,
-		step: sc?.step,
-	};
 }
 
 describe(Point3dInputPlugin.id, () => {
